Added solutions to 4.15-4.22

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogRouter.get("/", async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   res.json(blogs);
 });
 
@@ -17,15 +19,28 @@ blogRouter.get("/:id", async (req, res) => {
 
 blogRouter.post("/", async (req, res) => {
   const body = req.body;
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({
+      error: "token missing or invalid"
+    });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({
+    url: body.url,
     title: body.title,
     author: body.author,
-    url: body.url,
+    user: user._id,
     likes: body.likes || 0
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   res.json(savedBlog);
 });
 
@@ -41,6 +56,23 @@ blogRouter.put("/:id", async (req, res) => {
 });
 
 blogRouter.delete("/:id", async (req, res) => {
+  const id = req.params.id;
+  const blogToDelete = await Blog.findById(id);
+
+  if (!blogToDelete) {
+    return res.status(400).json({
+      error: `No blog with the following id: ${id}`
+    });
+  }
+
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: "token missing or invalid" });
+  } else if (decodedToken.id !== blogToDelete.user.toString()) {
+    return res.status(401).json({ error: "Not authorized" });
+  }
+
   await Blog.findByIdAndRemove(req.params.id);
   res.status(204).end();
 });
